requirements.txt
absl-py==2.3.0
ale-py==0.11.0
asttokens==3.0.0
captum==0.8.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
imageio==2.37.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pip==24.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.0
psutil==5.9.5
pure_eval==0.2.3
pyboy==2.4.0
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
PySDL2==0.9.17
pysdl2-dll==2.32.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==304
pyzmq==25.0.2
rich==14.0.0
sb3_contrib==2.6.0
scikit-image==0.25.2
scipy==1.15.3
setuptools==72.1.0
six==1.16.0
stable_baselines3==2.6.0
stack-data==0.6.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
tornado==6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.43.0
zipp==3.20.2
